refactoring: make People.generatePairsExcluding() accept excluded pairs as method argument
Before we initialized People object with excluded pairs before pairs generation
using method excludePairs(). Analyzing its usage we could see that it was
always followed by generatePairs() method call.

There is small sense in keeping expluded pairs inside of People objects,
as it is used only in in context of pairs generation, and between
different round excluded pairs will change and should be updated each time,
so we need to remember to do it before next generatePairs() call.

As this is error prone and inconvenient it's better to pass it as an argument
to People.generatePairs() method. As we already have method with same name
code was refactored to introduce new generatePairsExcluding() method
which accepts excluded pair as an input. Existing generatePairs() method
was reworked to use new method with empty excluded pairs list.